Added an example line to disable resource checking
Added an example line to disable resource checking so it can be deployed in an small VM for testing purposes.
The line was added commented so the installer can remind the user the recommended values. Having it there is more discoverable for users, reducing the time and effort to make it run.