Set the default ES memory limit to 4Gi
This patch sets the default ES memory limit to 4Gi in our vars.yml
template.  The defalut in openshift-ansible is 8Gi, which can result
in the ES container never being deployed for AIO deployments on
systems with smaller amounts of memory.  Our ViaQ inventory files
set openshift_check_min_host_memory_gb=7, but ES fails to deploy
on systems with 10G of memory unless the memory limit is reduced.